Moving duplciate code into PostList
