+++ src/Home.tsx
@@ -1,16 +1,9 @@
-import { ActionPanel, Action, Icon, List, showToast, Toast } from "@raycast/api";
-import { AbortError } from "node-fetch";
-import { useEffect, useRef, useState } from "react";
-import { searchAll } from "./RedditApi/Api";
-import RedditResultItem from "./RedditApi/RedditResultItem";
+import { ActionPanel, Action, Icon, List } from "@raycast/api";
+import { useState } from "react";
 import { joinWithBaseUrl } from "./RedditApi/UrlBuilder";
 import SubredditList from "./SubredditList";
 import FilterBySubredditPostList from "./FilterBySubredditPostList";
-import Sort from "./Sort";
-import RedditSort from "./RedditSort";
-
 import PostList from "./PostList";
-import getPreferences from "./Preferences";
 
 export default function Home({
   favorites,
@@ -21,72 +14,12 @@ export default function Home({
   addFavoriteSubreddit: (subreddit: string) => void;
   removeFavoriteSubreddit: (subreddit: string) => void;
 }) {
-  const [results, setResults] = useState<RedditResultItem[]>([]);
   const [searching, setSearching] = useState(false);
-  const [searchRedditUrl, setSearchRedditUrl] = useState("");
-  const [sort, setSort] = useState(RedditSort.relevance);
-  const abortControllerRef = useRef<AbortController | null>(null);
-  const queryRef = useRef<string>("");
-
-  const doSearch = async (query: string, sort = RedditSort.relevance, after = "") => {
-    abortControllerRef.current?.abort();
-    abortControllerRef.current = new AbortController();
-
-    setSearching(true);
-    if (!after) {
-      setResults([]);
-    }
-
-    setSort(sort);
-    queryRef.current = query;
-
-    if (!query) {
-      setSearching(false);
-      return;
-    }
-
-    try {
-      const preferences = getPreferences();
-      const apiResults = await searchAll(
-        "",
-        query,
-        preferences.resultLimit,
-        sort?.sortValue ?? "",
-        after,
-        abortControllerRef.current
-      );
-      setSearchRedditUrl(apiResults.url);
-
-      if (after) {
-        setResults([...results, ...apiResults.items]);
-      } else {
-        setResults(apiResults.items);
-      }
-    } catch (error) {
-      if (error instanceof AbortError) {
-        return;
-      }
-
-      console.log(error);
-      showToast({
-        style: Toast.Style.Failure,
-        title: "Something went wrong :(",
-        message: String(error),
-      });
-    } finally {
-      setSearching(false);
-    }
-  };
-
-  useEffect(() => {
-    return () => {
-      abortControllerRef?.current?.abort();
-    };
-  }, []);
+  const [query, setQuery] = useState<string>("");
 
   return (
-    <List isLoading={searching} onSearchTextChange={doSearch} throttle searchBarPlaceholder="Search Reddit...">
-      {!queryRef.current && (
+    <List isLoading={searching} onSearchTextChange={setQuery} throttle searchBarPlaceholder="Search Reddit...">
+      {!query && (
         <>
           <List.Section title="More ways to search">
             <List.Item
@@ -135,12 +68,7 @@ export default function Home({
           </List.Section>
         </>
       )}
-      <PostList
-        posts={results}
-        sort={sort}
-        searchRedditUrl={searchRedditUrl}
-        doSearch={(sort: Sort, after = "") => doSearch(queryRef.current, sort, after)}
-      />
+      <PostList setSearching={setSearching} query={query} />
     </List>
   );
 }
