+++ src/PostList.tsx
@@ -1,29 +1,93 @@
-import { ActionPanel, Action, Icon, List } from "@raycast/api";
+import { ActionPanel, Action, Icon, List, showToast, Toast } from "@raycast/api";
 import RedditResultItem from "./RedditApi/RedditResultItem";
 import PostActionPanel from "./PostActionPanel";
-import Sort from "./Sort";
 import redditSort from "./RedditSort";
 import SortListItem from "./SortListItem";
+import RedditSort from "./RedditSort";
+import { useEffect, useRef, useState } from "react";
+import getPreferences from "./Preferences";
+import { searchAll } from "./RedditApi/Api";
+import { AbortError } from "node-fetch";
 
 export default function PostList({
-  posts,
-  sort,
-  searchRedditUrl,
-  doSearch,
+  setSearching,
+  subreddit = "",
+  query,
 }: {
-  posts: RedditResultItem[];
-  sort: Sort;
-  searchRedditUrl: string;
-  doSearch: (sort: Sort, after?: string) => void;
+  setSearching: (searching: boolean) => void;
+  subreddit?: string;
+  query: string;
 }) {
-  if (!posts.length) {
+  const abortControllerRef = useRef<AbortController | null>(null);
+  const [results, setResults] = useState<RedditResultItem[]>([]);
+  const [sort, setSort] = useState(RedditSort.relevance);
+  const [searchRedditUrl, setSearchRedditUrl] = useState("");
+
+  const doSearch = async (sort = RedditSort.relevance, after = "") => {
+    abortControllerRef.current?.abort();
+    abortControllerRef.current = new AbortController();
+
+    setSearching(true);
+    if (!after) {
+      setResults([]);
+    }
+
+    setSort(sort);
+
+    if (!query) {
+      setSearching(false);
+      return;
+    }
+
+    try {
+      const preferences = getPreferences();
+      const apiResults = await searchAll(
+        subreddit,
+        query,
+        preferences.resultLimit,
+        sort?.sortValue ?? "",
+        after,
+        abortControllerRef.current
+      );
+      setSearchRedditUrl(apiResults.url);
+
+      if (after) {
+        setResults([...results, ...apiResults.items]);
+      } else {
+        setResults(apiResults.items);
+      }
+    } catch (error) {
+      if (error instanceof AbortError) {
+        return;
+      }
+
+      console.log(error);
+      showToast({
+        style: Toast.Style.Failure,
+        title: "Something went wrong :(",
+        message: String(error),
+      });
+    } finally {
+      setSearching(false);
+    }
+  };
+
+  useEffect(() => {
+    doSearch();
+
+    return () => {
+      abortControllerRef?.current?.abort();
+    };
+  }, [query]);
+
+  if (!results.length) {
     return null;
   }
 
   return (
     <>
       <List.Section title={sort ? `Results (Sorted by ${sort.name})` : "Results"}>
-        {posts.map((x) => (
+        {results.map((x) => (
           <List.Item
             key={x.id}
             icon={
@@ -42,7 +106,7 @@ export default function PostList({
           title="Show more..."
           actions={
             <ActionPanel>
-              <Action title="Show more..." onAction={() => doSearch(sort, posts[posts.length - 1].afterId)} />
+              <Action title="Show more..." onAction={() => doSearch(sort, results[results.length - 1].afterId)} />
             </ActionPanel>
           }
         />
